west.yml: update nxp_hal SHA to latest commit
Commit 5852d8f from https://github.com/zephyrproject-rtos/hal_nxp
fixed the import script, but there's no corresponding update for the
west.yml configuration file. As such, if one attempts to add support
for new drivers, they'll be copied in the wrong places
(under $ZEPHYR_BASE). As such, modify the west.yml file to point to
the HEAD of the master branch of the hal_nxp repo.

Fixes: 5852d8f mcux: adjust paths in import script after moving
               HAL to module

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>