Add CONFIG_GREYBUS_STATIC_MANIFEST
For very specific devices, it would be advantageous to compile-in a static, constant precompiled Greybus Manifest.

Initially, this will take the form of the example include/greybus/static-manifest.h, where the encoding is done
with e.g. xxd and then manually edited to add the "static const" prefix and some guard headers.

Eventually, the CONFIG_GREYBUS_STATIC_MANIFEST should point to a '.mnfs' file (a Greybus Manifest *Source* file),
the build system will call the 'manifesto' tool to parse it and compile it to a '.mnfb' file (a Greybus Manifest *Binary* file), and then
the build system will call xxd through some script to put the output in the expected format.