Locate and Plug Memory Leak in Greybus
The root cause of this memory leak was actually due to different semantics of the function

atomic_val_t atomic_dec(atomic_t *target);

between RTOS's.

In NuttX (ant the equivalent call in Linux), atomic_dec() returns the value of target *after* it has been decremented.

However, in Zephyr (and also in stdatomic.h / atomic_fetch_sub()) the value of target *before* it is decremented is returned.

That difference in behaviour meant that greybus operations were never being destroyed because the returned reference count was only going as low as 1 and it was only ever decremented once.

A simple 1-line change corrects the behaviour.