subsys: greybus: initial support for Greybus in Zephyr
Greybus was originally designed for the Project Ara
modular smartphone and has since been merged in the
upstream Linux kernel. Some parts are still in staging.

We are leveraging some of that work to pursue Greybus
as a potential community-driven application layer
protocol for IoT.

This Greybus implementation was adapted from the
Project Ara reference design, which originally used
Nuttx in collaboration with BeagleBoard.org.

Thanks to everyone involved.

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>