pthread: facilitate dynamically assigned / allocated stacks
This change allows users to pre-allocate a configurable number of POSIX
thread stacks via Kconfig. It also facilitates a future implementation
where stacks may be heap-allocated on demand or a hybrid approach. Of
course, user-allocated stacks continue to work as usual. Thus, allowing
both a NULL and non-NULL pthread_attr_t to be supplied to
pthread_create(3) and being more standards-compliant.

Fixes #25973

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>