greybus: platform: spi: populate num_chipselect from dt info
The Greybus SPI controller (a.k.a. master) node in DT is of compat
"zephyr,greybus-spi-controller" whereas the peripherals (a.k.a. slaves)
attached to the bus are of compat "zephyr,greybus-spi-peripheral".

This change automatically populates the num_chipselect field of
the gb_spi_master_config_response for the SPI controller to reflect the
number of attached peripherals that are of compat
"zephyr,greybus-spi-peripheral" and have status "okay".

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>