greybus: transport: uart: add uart transport and update sample
This change brings samples/subsys/greybus/uart a little closer
to being up to speed with samples/subsys/greybus/net.

There seems to be some miscommunication somewhere though because
there are timeouts soon after the control connection is established.

Since uart is not the primary interest right now, it's probably
safe to get this change in and fix whatever the uart bug is at
a later time. The uart has always been a bit flaky with gbridge
anyway.

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>