known-issues: delete out of date content
It turns out that current known issues folder contains issues that are
no longer valid or present. I have searched for some of the regex
snippets listed but I have not been able to find any. Documentation was
the last active user of the known-issues folder, but has moved to
another solution.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>