boards: hexiwear_kw40z: Remove I2C pins and dts enablement
I2C isn't connected on the KW40Z to anything so configuring the pins
for at and having it enabled in the devicetree don't make any sense.

Additionally, the pins for I2C are currently conflicting with the UART
pins, which is another reason to remove the I2C config.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>