dts: bindings: sx12xx: add ST stm32wl-subghz-radio node
Add a dts node for st,stm32wl-subghz-radio. The device is integrated in
STM32WL series SoCs, and based on the sx1262, but it does not use any
gpio as all the necessary signals are internally connected to various
SoC units.

To account for that, make the redundant gpio optional in the template
files, but mark them as required in the sx1261 and sx1262 definitions,
to match with what's used in the driver.

Signed-off-by: Fabio Baltieri <fabio.baltieri@gmail.com>