tests: lib: cmsis_dsp: matrix: Run binary_q15 tests on mps2_an521
The CMSIS-DSP matrix binary_q15 tests should not be excluded on the
mps2_an521 board, which is one of the integration platforms.

This test was likely excluded because of the computation precision-
related failures; but, this failure is no longer observed, so there is
no reason to exclude it anymore.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>