api: gpio: add gpio direction query functions and driver api
Some services require the ability to query whether a GPIO pin
is configured as an input, output, both, or neither prior to
performing any operations at the service level. This is done
in order to reduce state tracking within the service.

To that end, this change adds
* `gpio_port_get_direction_bits()`
* `gpio_pin_is_input()`, and
* `gpio_pin_is_output()`

as well as `CONFIG_GPIO_GET_DIRECTION_BITS` so that drivers
may enable the API if it is supported.

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>